refactor: 로컬 인증 기능 패키지 분리 (#49)
- AuthController:
  - 일반 로그인과 회원가입 API를 local 패키지로 이동
  - 기존 API 경로와 응답 규격 유지
- AuthService:
  - 회원 조회와 등록을 회원 서비스 경계로 위임
- request/response DTO:
  - 로컬 인증 전용 패키지로 이동
