--- src/main/java/org/zerock/puppyrun/auth/controller/response/SignInResponse.java
@@ -1,10 +1,0 @@
-package org.zerock.puppyrun.auth.controller.response;
-
-import lombok.Builder;
-
-@Builder
-public record SignInResponse(
-        String accessToken,
-        String refreshToken
-) {
-}
