+++ src/main/java/org/zerock/puppyrun/auth/local/controller/request/SignInRequest.java
@@ -1,4 +1,4 @@
-package org.zerock.puppyrun.auth.controller.request;
+package org.zerock.puppyrun.auth.local.controller.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
