+++ src/main/java/org/zerock/puppyrun/auth/local/controller/response/CheckResponse.java
@@ -1,4 +1,4 @@
-package org.zerock.puppyrun.auth.controller.response;
+package org.zerock.puppyrun.auth.local.controller.response;
 
 import lombok.Builder;
