+++ src/main/java/org/zerock/puppyrun/auth/local/controller/response/SignInResponse.java
@@ -1,9 +1,9 @@
-package org.zerock.puppyrun.auth.controller.response;
+package org.zerock.puppyrun.auth.local.controller.response;
 
 import lombok.Builder;
 
 @Builder
-public record TokenReissuanceResponse(
+public record SignInResponse(
         String accessToken,
         String refreshToken
 ) {
