+++ src/main/java/org/zerock/puppyrun/auth/local/controller/response/TokenReissuanceResponse.java
@@ -1,0 +1,10 @@
+package org.zerock.puppyrun.auth.local.controller.response;
+
+import lombok.Builder;
+
+@Builder
+public record TokenReissuanceResponse(
+        String accessToken,
+        String refreshToken
+) {
+}
