+++ src/main/java/org/zerock/puppyrun/auth/local/service/AuthService.java
@@ -1,4 +1,4 @@
-package org.zerock.puppyrun.auth.service;
+package org.zerock.puppyrun.auth.local.service;
 
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -8,43 +8,43 @@ import org.springframework.transaction.annotation.Transactional;
 import org.zerock.puppyrun.common.auth.jwt.JwtTokenProvider;
 import org.zerock.puppyrun.member.DTO.MemberDTO;
 import org.zerock.puppyrun.auth.DTO.TokenDTO;
-import org.zerock.puppyrun.auth.controller.request.SignInRequest;
-import org.zerock.puppyrun.auth.controller.request.SignUpRequest;
-import org.zerock.puppyrun.member.entity.Member;
+import org.zerock.puppyrun.auth.local.controller.request.SignInRequest;
+import org.zerock.puppyrun.auth.local.controller.request.SignUpRequest;
 import org.zerock.puppyrun.member.exception.ExistingUserException;
 import org.zerock.puppyrun.member.exception.UserNotFoundException;
 import org.zerock.puppyrun.member.exception.UserUnauthorizedException;
-import org.zerock.puppyrun.member.repository.MemberRepository;
+import org.zerock.puppyrun.member.service.MemberQueryService;
+import org.zerock.puppyrun.member.service.MemberRegistrationService;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class AuthService {
-    private final MemberRepository memberRepository;
+    private final MemberQueryService memberQueryService;
+    private final MemberRegistrationService memberRegistrationService;
     private final PasswordEncoder passwordEncoder;
     private final JwtTokenProvider jwtTokenProvider;
 
 
     // 닉네임 중복 검증
     public boolean isExistsByNickname(String nickName) {
-        return memberRepository.existsByNickName(nickName);
+        return memberQueryService.existsByNickname(nickName);
     }
 
     // 이메일 중복 검증
     public boolean isExistsByEmail(String email) {
-        return memberRepository.existsByEmail(email);
+        return memberQueryService.existsByEmail(email);
     }
 
-    protected Member findMemberById(UUID id) {
-        return memberRepository.findById(id)
-                .orElseThrow(() -> new UserNotFoundException("존재하지 않는 회원입니다."));
+    protected MemberDTO findMemberById(UUID id) {
+        return memberQueryService.findByIdOrThrow(id);
     }
 
     /**
      * AccessToken 및 RefreshToken 생성
      *
-     * @param MemberDTO
-     * @return TokenDTO
+     * @param memberDTO 토큰에 포함할 회원 정보
+     * @return 액세스 토큰과 리프레시 토큰
      */
     public TokenDTO crateToken(MemberDTO memberDTO) {
         return TokenDTO.builder()
@@ -56,8 +56,8 @@ public class AuthService {
     /**
      * 회원가입 처리 메서드
      *
-     * @param SignUpRequest
-     * @return MemberDTO
+     * @param request 회원가입 요청
+     * @return 생성된 회원 정보
      */
     @Transactional
     public MemberDTO registrarMember(SignUpRequest request) {
@@ -71,35 +71,28 @@ public class AuthService {
         // 비밀번호 암호화
         String encryptedPassword = passwordEncoder.encode(request.password());
 
-        Member member = Member.builder()
-                .email(request.email())
-                .nickName(request.nickName())
-                .password(encryptedPassword)
-                .build();
-
-        Member savedMember = memberRepository.save(member);
-        return savedMember.toDto();
+        return memberRegistrationService.registerLocalMember(
+                request.email(),
+                request.nickName(),
+                encryptedPassword
+        );
     }
 
     /**
      * 로그인 처리 (AccessToken,RefreshToken 포함)
      *
-     * @param SignInRequest
-     * @return TokenDTO
+     * @param request 이메일과 비밀번호를 포함한 로그인 요청
+     * @return 액세스 토큰과 리프레시 토큰
      */
     public TokenDTO signIn(SignInRequest request) {
-        return memberRepository.findByEmail(request.email())
-                .map(member -> {
+        return memberQueryService.findCredentialsByEmail(request.email())
+                .map(credentials -> {
                     // 비밀번호 검증
-                    if (!passwordEncoder.matches(request.password(), member.getPassword())) {
+                    if (!passwordEncoder.matches(request.password(), credentials.encodedPassword())) {
                         throw new UserUnauthorizedException("비밀번호가 틀립니다.");
                     }
-
                     // 토큰 쌍 생성 (Access + Refresh)
-                    return TokenDTO.builder()
-                            .accessToken(jwtTokenProvider.generateAccessToken(member.toDto()))
-                            .refreshToken(jwtTokenProvider.generateRefreshToken(member.toDto()))
-                            .build();
+                    return crateToken(credentials.member());
                 })
                 .orElseThrow(() -> new UserNotFoundException("존재하지 않는 회원입니다."));
     }
@@ -107,21 +100,17 @@ public class AuthService {
     /**
      * AccessToken 재발급
      *
-     * @return TokenDTO
+     * @param refreshToken 사용자 식별에 사용할 리프레시 토큰
+     * @return 새 액세스 토큰과 리프레시 토큰
      */
     public TokenDTO AccessTokenReissuance(String refreshToken) {
         UUID userId = jwtTokenProvider.getUserIdFromRefreshToken(refreshToken);
-
-        Member member = findMemberById(userId);
-
-        String newAccessToken = jwtTokenProvider.generateAccessToken(member.toDto());
-        String newRefreshToken = jwtTokenProvider.generateRefreshToken(member.toDto());
+        MemberDTO member = findMemberById(userId);
 
         return TokenDTO.builder()
-                .accessToken(newAccessToken)
-                .refreshToken(newRefreshToken) // 새 리프레시 토큰 반환
+                .accessToken(jwtTokenProvider.generateAccessToken(member))
+                .refreshToken(jwtTokenProvider.generateRefreshToken(member))
                 .build();
     }
-
 
 }
